LeAudio: Fix state machine handling ACL disconnection
When the state machine is handling the disconnection state,
the device for a while still has the valid connection ID and
state. Configuration invalidation logic thinks the group still
has both devices connected and does not select the proper
configuration. This change moves the configuraiton cache
refresh after the connecion ID of the device is set to invalid.

Bug: 275154325
Test: atest --host bluetooth_le_audio_client_test bluetooth_le_audio_test --no-bazel-mode
Change-Id: Ie8a2147d7759defdb21e15a82a5a858386957d47